[kernel][test] Add test exercising singleton mutexes.
The test ensures that acquiring a singleton mutex is thread-safe the
first time it is acquired.

At time of writing, we have a bug where singleton mutexes are lazily
initialised, but not initialised in a thread-safe manner. This means
that the first time a mutex it is possible that two threads may hold it
simultaneously.

This test sets up N threads and races them acquiring a singleton mutex
While the test is safe to run multiple times, it can only exercise the
static initialisation code path once per boot.

Bug: 78920
Change-Id: I28d6eb90bb46e681297f59391207b78792ee5626
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/544202
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>