[input] Validate the well-formedness of KeyEvent
The commit at fxr/573576 started enforcing timestamps on `KeyEvent`s
forwarded by the IME service to clients. This was an improvement over
prior state of the art but didn't go far enough.

This change adds verification on the fields `type_`, `key` and
`key_meaning`.  If we forward `KeyEvents` that don't have them this is a
bug, so we crash with an appropriate error message.  I don't think it is
possible to make this check happen at compile time.

Bug: 83632
Change-Id: I5e9e10610f08063f9c7a06b02b3128dea01c3702
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575246
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>