[setui][tests] Allow finer controlled trigger
This CL allows the light sensor scanning code to use a generic trigger
for starting processing of the next light sensor event. It moves the
trigger creation to the caller, so the production code can still trigger
a time based event, but the tests can use a manual trigger to be more
deterministic in the evaluation of the code. This should hopefully
prevent test flakes due to timeouts we've been seeing.

Fixed: 82610
Test: fx test -o setui_service_tests
MULTIPLY: setui_service_tests:10
Change-Id: Iae35e8795b91599641754e173f5e6a15b62c8b45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573421
Commit-Queue: Paul Faria <paulfaria@google.com>
Reviewed-by: Sijie Chen <sijiec@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>