[inspect][contrib] Workaround rustc issue
This CL updates the syntax in a macro call to avoid a bug in the rust
compiler that prevents us from analyzing nightly rolls in the rust
compiler. The issue itself is actively being worked on at
https://github.com/rust-lang/rust/issues/88307. While normally we would
not add a workaround for such issues, it's become too complicated to revert
the original change due to commits applied on top of the original
change. We've also been unable to analyze any other potential failures
that might be coming in. Given the active work on the issue, we felt it
was safe to add in this workaround so we could track any other possible
issues coming in during this time.

Fixed: 83801
Change-Id: Ie7aa6128e3f0eda9aad7bc5f29072f84feb77185
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575604
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>