[modular] Fix use-after-free in StoryControllerImpl
StoryStorage may outlive StoryControllerImpl because it's owned by
SessionStorage and not destroyed together with StoryControllerImpl.

It's possible for StoryStorage to be updated and invoke this listener
after the StoryControllerImpl is destroyed, for example when an
AddModCall is processed during teardown.

Fixed: 83815
Change-Id: I43ec1b26ca09b10347a2673ef203cdbe9e668b4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575746
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Fuchsia-Auto-Submit: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>