[kernel][thread] Change assert in SetMigrateFnLocked
When a VCPU is being destroyed, we need to be able to clear the
migration function, otherwise the scheduler may call the migration
function and incorrectly refer to freed memory.

It should be safe to set the migration function to nullptr, even if we
are in the midst of migrating the thread.

Change-Id: I5f0517fc18d39d247fbf498ffeb2b3d2b5c36926
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574621
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>