[cm_config] Add builtin_capabilities field to RuntimeConfig
Currently, component manager will route any capability that
claims to be a built-in capability. (It only then opens the
capability if it's actually implemented as a built-in.) In
order to statically verify capability routes, we need some
way to specify the intended set of built-ins in a config or
manifest file. This change adds a Vec<CapabilityDecl>-valued
field to the runtime config.

In a follow-up, component manager's routing code (and the
static routing analyzer) will reject any uses of
capabilities that are routed as built-ins but not registered
in this config field.

Bug: 83626

Change-Id: I8ca30ec374e8925b2521dfb3661c1c913faac17a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574148
Fuchsia-Auto-Submit: Laura Peskin <pesk@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>