[wlan][iwlwifi] Minor logic change for SimNvm::HandleChunkRead()
A follow-up of 9ec5d8fbc4. The original code breaks the loop and
returns an empty data at end of function when the 'data' field is
empty. This CL behaves the same, but changes the logic a little bit.

Bug: 46688
Tested: Verified in local test and CQ.

Change-Id: I02d5c58949254bbe25520088f4fb1596e99b479a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575281
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>