[usb-virtual-bus-launcher] Fix so that zxtest assertion failures aren't ignored
Before this change, if I add EXPECT_TRUE(0) somewhere in
USBVirtualBusBase() in
zircon/system/ulib/usb-virtual-bus-launcher/usb-virtual-bus-launcher.cc
(or add ASSERT_TRUE(0) at the end of that function), an assertion
failure message gets printed in the test output, but the test passes
overall.

That happens because this code is linked into a dynamic library and
links against zxtest, so it gets a separate copy of the zxtest library
(which is currently statically linked) and hence has a different
instance of the global variable that tracks whether an assertion
failure has occurred.  The copy of zxtest used by main() therefore
doesn't see the failure.  zxtest currently isn't safe to be used from
dynamic libraries this way.

Fix this by statically linking this code.

Bug: 83770
Test: see above
Change-Id: I93b026ba634b9e3e57238536019c6cbafe1dc966
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575792
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>