[fidlc] Add attribute for overriding anonymous layout name
See bug for context. This attribute is placed on the layout
directly rather than on the member, since it does not
affect the anonymous naming scope (introduced by the
member name), just the generated name itself. This allows the
result union special name handling to be reimplemented
in terms of the @generated_name attribute.

Implementation is straightforward but for one aspect:
the override name is required for compilation, but we
don't have access to the resolved Constant value of
the attribute arg before compilation. To get around this
we use the span of the attribute arg itself, and do
validation of the argument/compilation later. This also
means that Attribute::GetArg returns a Constant rather
than a ConstantValue, so that we have access to the
attribute arg's span.

Test: fx ninja -C out/default host_x64/fidl-compiler &&
./out/default/host_x64/fidl-compiler

Fixed: 81550
Change-Id: If77e70c121a96ce28cadcd5347084f95ad9b6341
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568581
Commit-Queue: Felix Zhu <fcz@google.com>
Reviewed-by: Alex Zaslavsky <azaslavsky@google.com>