[cm][runtime_config] Register builtin capabilities
In a follow-up, component manager's router will reject
uses of capabilities that are routed as built-ins but
are not registered in builtin_capabilities field of the
runtime config.

This change registers the builtin capabilities that
are used in real and integration-test instances of
component framework.

Bug: 83626

Change-Id: I7b879d1c22315c7f8eca576a4516b3521bb9b0fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/556667
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Fuchsia-Auto-Submit: Laura Peskin <pesk@google.com>
Commit-Queue: Laura Peskin <pesk@google.com>