[identity] Add Identity API methods to support authn prototype.
This is the second in a chain of CLs to adapt the existing identity API
surface to the simple authentication prototype that we're in the process
of building.

Step 2 is to add in temporary methods (marked as deprecated) to get and
to create an account under the hardcoded assumption that authentication
is password. This also adds an additional longer term method to get
account metadata from the account manager interface.

Bug: 81605
Change-Id: I443f4f667068727a5d202ad8f53e932bf3de66db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573578
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>