[kernel][crashlog] Always finalize crash logs
Always finalize crashlogs, even when no crashlog body is rendered.
This makes sure that the even when backtraces and panic buffers are
not going to be included in the log, that the crash reason is properly
reported, and that EFI based crashlog systems have a chance to perform
any required stowing actions.

Bug: 83742
Tests: new integration tests
Change-Id: I9c9f54145225bb78ffbd7451379e0c6026261e3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575789
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Commit-Queue: John Grossman <johngro@google.com>