[connectivity][fuzzing] Remove zxtest dep from usb-cdc-ecm fuzzer.
The usb-cdc-ecm fuzzer had a GN dep on zxtest. It is unneeded, but
didn't harm anything for fuzzers running as v1 component. For v2
components, it introduces a dependency on the elf test runner that
conflicts with the generated fuzzer unit test's dependency on the gtest
test runner.

Bug: 76528
Change-Id: I08a717a95c23f0b212ccda1c0c11d11539464816
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573572
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>