[kernel][vm] Support for eviction hints
Add support for eviction hints in VmObjectPaged. There are two hints
that can be indicated:
 - AlwaysNeed: The range is always needed and should not be considered
 for memory reclamation (unless the kernel decides otherwise, e.g. to
 prevent an OOM reboot). Any absent pages in the range are committed
 first.
 - DontNeed: The range is not needed anymore and can be marked
 "inactive", moving the pages to the front of the line for reclamation.
 This only applies to pages already present in the range; no new pages
 are committed.

Hints only apply to pager-backed VMOs for now, and are no-ops for other
VMO types.

This commit only puts in the internal VMO machinery. The API is not
exposed to userspace yet. Syscalls that exercise this will follow in
separate commits.

Test: kernel-unittests

Bug: 65334
Change-Id: Idc98e2572a10a4bae66c8f6a15d18ae8b94ac36c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566022
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>