[ftln] Remove unused data scan in spare validity calculation.
The original intention was to check for anything that appears to be
non-erase bytes in the data section of the ECC page as well, and that
information was scanned for but erroneously unused. Experience so far is
that we have always had partial writes on the full page boundary, but
even in the case of partial writes on the ECC page boundary this should
still catch the issue since the spare bytes are spread out across the
ECC pages.

Bug: 83474
Change-Id: Iaf2ae4398aaca46d3504166362b2f397d7b13790
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575748
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>