[phys] Clarify semantics of BootZbi::Load()
It is 'morally' the case today that the associated kernel image of a
BootZbi is loadable in-place after a successful Load(); however, the
class's internal bookkeeping does not reliably recognize that. This CL
fixes that, begins enforcing the invariant with a debug assert in
Load(), and simplifies physboot's interaction with BootZbi in relying on
that invariant. That simplification of physboot also removes the buggy
assertions and expectations that led to fxbug.dev/81726.

Bug: 68585
Fixed: 81726
Change-Id: I7305cdb7c24f7fd1011840e1406747cefc89a20b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574782
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>