[fidl] Attributes support multiple arguments
FIDL versioning requires attributes that support multiple arguments,
like `@available(added=3,deprecated=4)`.  This change implements that
feature for both official "schema-ed" attributes (ie, those that are
officially support by the FIDL language, and thus listed at
https://fuchsia.dev/fuchsia-src/reference/fidl/language/attributes)
and user-defined "schema-less" attributes.  The former feature
support for verifying the presence of required arguments, while the
latter do not.

One interesting aspect of this change is how it deals with attributes
that carry "inferred" argument names, like `@foo("bar")`.  In the case
of official attributes, the compiler uses the schema to infer the name
of the lone argument.  Conversely, for unofficial attributes, this
name is always defaulted to "value."

In light of these changes, a number of small modifications have been
made to the schemas of various attributes:

*   `@doc`'s lone (required) argument is now called `"text"`.
*   `@transport`'s lone (required) argument is now called `"types"`.
*   `@transitional`'s lone (optional) argument is now called,
    `"reason"`.

Fixed: 78236
Test: fidldev test fidlc
Change-Id: I69df67807c7819c1258f69e525736084f9d93698
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/560962
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>