[brightness] Remove frequent service startup logging
Workstation shell connects and disconnects from various settings related
services when displaying sysui overlays. This is to prevent background
load impose by sysui when it is not visible. Every time the sysui is
displayed, it connects to the settings services, in this case brightness
service, and cause this log to be displayed.

In a typical usage, for example in workstation build walkthroughs, the
sysui will be displayed tens of times within a minute. Having this log
is not very helpful and not consist with other setting services.

Change-Id: Ia3d2745f1c32905ac1f1d41befe3cd6bdccc547b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576221
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Fuchsia-Auto-Submit: Kevin Twidle <kpt@google.com>
Reviewed-by: Kevin Twidle <kpt@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>