[display][i915] Prevent client GTT allocations over framebuffer
The bootloader framebuffer is located at the base of the "data stolen
memory" region. When initializing the address mapping in the global GTT,
the driver initializes the region allocator in a way that allows clients
of fuchsia.hardware.intelgpucore to modify the contents of the
framebuffer directly.

This is now prevented by excluding this region from the GTT region
allocator.

Bug: 80350
Change-Id: Ic6ca18a8aee1e263c15ef9be9cb11c112a9a13c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/566901
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: John Bauman <jbauman@google.com>