[swd][security] Remove SWD policy target references in core.gni
This change removes the explicitly definition of SWD policy targets in
core.gni and instead relies solely on the definition of
policy_labels.swd to determine which SWD policy targets will be included
in the final image assembly.

This was tested by inspecting the built images/blobfs artifacts to
ensure that the expected targets were properly applied; specifically the
existence/non-existence of the disable-executability-restrictions file
and specific pkg-resolver config-data files.
This was done on the following builds:
- core.x64 (core.gni)

For reference, the migration process is as follows:

1. Define all SWD policies.
2. Add dependency on SWD policies in base_package_deps and
   system_image_deps, but do not apply a SWD policy to retain existing
   product behavior.
3. Apply SWD policies in product definitions alongside the existing
   SWD targets configuration. Assert that product configuration has not
   changed.
4. Remove manually defined SWD targets in product definitions, rely
   solely on defined SWD policies.

Change-Id: I77817e98160a43be5f88bfd7116ff35dde328ae2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575442
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Daniel Hwang <arkay@google.com>