[pwrbtn-monitor] Support setting action to perform on power button.
On Chromebooks, the power button is used by the TPM to verify that a
user is present at the device. This is necessary when enabling
case-closed debugging or when using the TPM as a U2F key. Before this
CL, pressing the power button in these scenarios would shut down the
system.

This CL adds an interface that allows components to configure the
behaviour of powerbtn-monitor so that the system isn't shut down in
scenarios like this.

Bug: 82504
Change-Id: Ie60919f8ac098808d6672b798a23e8f0c98027e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573203
API-Review: Alice Neels <neelsa@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>