[realm-builder] place services in the v1 realm
Provide v1 components launched by realm builder with their services by
providing them in the `additional_services` field of the singleton realm
that is constructed instead of by providing them to the component
itself. This means that any other v1 components that are manually
launched in that realm will be able to access these services as well,
which is more in line with common expectations in v1.

Change-Id: If82d7187c088d93de1856ab2a2fcb7e4843e7e2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576102
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>