[safemath] Replace zx_library() with library_headers().
The safemath header-only library can be modeled by a library_headers()
target instead of a zx_library() one. This avoids the creation of an
empty libsafemath.a file that is linked into all dependents, without
changing the semantics of the target.

This is the only use of 'kernel = false' in a zx_library() instance,
which is replaced here with an explicit assert. This CL thus prepares
for the removal of the 'kernel' parameter from zx_library() in a
future CL.

Bug: 83881
Change-Id: I72bee4d1aa71a216660c6b977db1843a2101e849
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576425
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Fang <fangism@google.com>