[kernel][crashlog] Always render a crashlog payload.
Always use crashlog_to_string to render a crashlog payload, regardless
of the reason for the reboot.  Allow crashlog_to_string to choose
which sections of a crashlog payload will be included based on the
reason for the crash.

Additionally, enable rendering of critical counters and recent debug
log info when crashing because of an OOM condition, and make root job
termination include the same information as an OOM reboot.

Graceful reboots continue to have no extra information included in
their crashlog payloads, however this should now be easy to change if
someone were inclined to do so.

Bug: 83742, 83845
Tests: updated integration tests.
Change-Id: I0df096bcd280592db648ca588ccf6109ca7121a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575943
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>