[codec_factory] Initial migration to v2
This change modifies Codec Factory so that it can run as a v1 or
a v2 component. The code looks at what services are available in
its environment to decide whether it is running as a v1 or v2
component. If running as a v1 component the Launcher service is
used. If running as v2 component the Realm service is used.

This change adds v2 test components which use the same code used
in v1 which validate operation of the software-based codecs.
Future changes will add more v2 tests.

This change does *not* move Codec Factory to v2 in production.

Bug: 78647

Change-Id: I351249b973b7e0d4705a3c730ee86f7e1d4863b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/563506
Reviewed-by: Ankur Mittal <anmittal@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>