[identity] Remove account transfer functionality
A couple of years ago we protoyped functions to transfer Fuchsia
accounts between devices using Overnet. The notion of a platform account
that spans devices is no longer in scope for Fuchsia so we should remove
this and the associated complexity from the API surface.

Account transfer was the only feature to use the "prototype mode" of
building account manager with features that weren't ready for prime
time, and structured config would be a better fit to meet this need if
it arises again in the future so this CL also removing the notion of
the account manager prototype.

This was pretty cool and I'm sad to see it go, but simpler is better.

Change-Id: If394dc86f8e81c36265bae6701f9da811aa37898
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576534
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>