[scrutiny] Hard fail on collection failures.
Since Scrutiny has moved to an in-memory data store for the model hard
faling on collectors makes more sense. Prior to this change if a
collector failed to collect required information Scrutiny would still
launch and work with an incomplete model. This was acceptable for when
the framework was used only as an auditing tool but as its scope has
expanded to being a compiled_action failures should occur as early in
the pipeline as possible. To achieve this I added a new
CollectorState::Failed which allows errors to be propagated and to abort
the running of scheduled collectors. This should greatly improve the
debugability of the framework.

Bug: 81806
Change-Id: Ifae3f832787a137b1d905b4d69907e879ac88845
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/564130
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>