[lsc] Start ScopedInstance via fuchsia.component.Binder
This CL proactively migrates certain components constructed via
ScopedInstance to connect to the Binder protocol of the instance's
exposed directory. It only does so for cases where no other connection is
made to the instance. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance. Currently, after a an instance is created via ScopedInstance,
it is automatically started. However, after the migration, this will no
longer be the case. As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400
Change-Id: If4f8b6bcb87ca6b300c8251fa71141d532ae06f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574410
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>