[lsc] Start realms via fuchsia.component.Binder
This CL proactively migrates certain realms constructed via
RealmBuilder to connect to the Binder protocol of the Realm's
root. It only does so for cases where no other connection is
made to the realm's children. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance, which RealmBuilder uses internally. Currently,
after a realm  is created via RealmBuilder, it is automatically started.
However, after the migration, this will no longer be the case.
As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400
Change-Id: I0994f304dd72e02830b98d9b5028753a9bfc2862
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574586
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Gary Bressler <geb@google.com>