[lsc] Start realms via fuchsia.component.Binder
This CL proactively migrates certain realms constructed via
RealmBuilder to connect to the Binder protocol of the Realm's
root. It only does so for cases where no other connection is
made to the realm's children. This is done to prepare for the
migration of `fuchsia.sys2.Realm/BindChild` to `fuchsia.sys2.Realm/OpenExposedDir`
in ScopedInstance, which RealmBuilder uses internally. Currently,
after a realm  is created via RealmBuilder, it is automatically started.
However, after the migration, this will no longer be the case.
As such, tests relying on this behavior will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400
Change-Id: Iabce8b0e0d100fb87a9bc558918d86c570ab5075
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574588
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>