[archivist] Block on component Start for test.
This CL changes the log::attribution::log_unattributed_stream
integration test for Archivist to block on start of test realm.
It does so to proactively prepare for a semantics change in the
RealmBuilder library.

A dependency of RealmBuilder, ScopedInstance, will migrate from
using `fuchsia.sys2.Realm/BindChild` to
`fuchsia.sys2.Realm/OpenExposedDir`. The primary consequence of
this change is that the instance created ScopedInstance won't
be automatically started. Since RealmBuilder creates its Realm
with this struct, it too will change.

In order to not break CI/CQ, and to keep this test green, the
test was changed to block on the Start event.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400
Change-Id: Id42b0d86b5eff879da240e0059d53d98be07c15e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574589
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>