[component_manager] Migrate ScopedInstance uses.
This CL proactively migrates uses of ScopedInstance in
test to connect to fuchsia.component.Binder of the exposed
directory of the instance. It only does so for cases where
no other connection is made to the instance. This is done
to prepare for the migration of `fuchsia.sys2.Realm/BindChild`
to `fuchsia.sys2.Realm/OpenExposedDir` in ScopedInstance.
Currently, after a child component is created via ScopedInstance,
it is automatically started. However, after the migration, this
will no longer be the case. As such, tests relying on this behavior
will break.

For more context on this migration, see
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0108_component_binder_protocol

Bug: 81401
Bug: 81400

Change-Id: If0e9c02a1c09ec782e6a4e0973a7bfb219315e46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/573464
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Casey Borders <casbor@google.com>
Reviewed-by: Gary Bressler <geb@google.com>