[kernel][hypervisor] Fix VMCS use-after-free
Ensures that if the user-space VCPU thread exits prior to the VCPU
object being destroyed that the VMCS is not left loaded. Previous the
VMCS could be left loaded on a CPU due to migrate fn clearing the
last_cpu_ field. With this field cleared the Vcpu destructor would not
know to perform vmclear, and the vmcs would remain loaded after its
page was returned to the pmm. At some future point the hardware could
write back to the VMCS, corrupting whatever had since allocated the
page.

This corruption could be easily observed by enabling the pmm checker
and running the core virtualization tests. With this change the pmm
checker no longer triggers.

Bug: 84002
Change-Id: Ie10e51ae4bd8872dad914a2932928a403e4c30a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577622
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>