[fxfs] Add pager support
This fills in the details so that we use the pager for reading from
cached objects. There are still some TODOs that will be addressed in
later CLs.

This is a reland and differs from the original by changing how locking
is handled around flushing. This version includes a lock that prevents
changes to the cache whilst take_flushable_data is called. There are
also some other less significant changes to the locks;
take_flushable_data no longer needs to return an event.

Change-Id: Iafc68d4ecc089ed5f57c218b78051d4b557dd3ba
Multiply: fxfs-append-tests 500
Multiply: fxfs-truncate-tests 500
Multiply: fxfs-random-op-tests 500
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576101
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>