[kernel][dlog] Fix dlog skipping empty newline
If a record contained only a newline character, it was removed and an
empty record was written to the serial without a newline character. This
CL updates that logic to not skip the only newline logs and to maintain
the contract that all logs end with a newline.

Bug: 83968
Bug: 83566
Test: k ut debuglog_tests

Change-Id: I2486f779964c46cf436619187826ecd3613c6a58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576909
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: John Grossman <johngro@google.com>