[exception][test]: Fix incorrect assertion.
zxtest assertions may only be used if zxtest::Run has been called.
Otherwise it may be used in the wrong context.

In this case, it is used from a child process that is meant to check
some other behavior.

Instead in order for the test to pass(return 0) the exception channel
creation must succeed.

Test: CQ
Bug: 83770
Change-Id: If2c24769755a69c41739564b37c404b3b3540059
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/575381
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>