[llcpp][fake-drivers] Copy radiobound bytes to std::vector
Currently, the fake-ot-radio test pushes a VectorView to the
radiobound_queue, but the lifetime of the VectorView's contents is only
up until the end of SendFrame(). This leads to a potential
use-after-free bug. Copying the bytes to an std::vector avoids this
issue.

Change-Id: Idfe9932527355258af1356218b320c67bfa83aa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578121
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>