[fvdl] Add dry-run flag and serde to ffx emu
In preparation for adding FVD metadata to ffx emu, we need to be able to
use a JSON file as input and quickly observe the effeects of that file
on the command's output.

This CL adds serde (rust (de)serialization to/from json) to the plugin's
args structure, and a dry-run flag to terminate the plugin as soon as
the command line is formed. This enables quick iteration times as we
work on defining the metadata schema for FVD going forward.

Bug: fxbug.dev/70828
Change-Id: I1407dc04ca38639b661612d90d99aa94f3f6afe2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577421
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>