[rust][rbe] Upload Rust standard libraries
This also includes a copy of libunwind.a from respective target lib dirs.
This is the library that is intended by this change, as it is now
requested by rustc linker commands (clang ... -lunwind).

This works around the issue where the underlying linker command (perhaps
unintentionally) requested -lunwind.

This might upload more stdlibs than are necessary for a given rust
compile, because previously the stdlibs were detected in the depfile
and individual uploaded.  This is probably ok because the stdlibs don't
change frequently.

See https://github.com/rust-lang/rust/pull/88483 for recently rolled
change.

Bug: 84162
Change-Id: I9204b81b843c646c190a9b13b288561d8cba5dd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578408
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: David Fang <fangism@google.com>