[virtualization] Modify termina_guest kernel/CIPD scripts
Update the "update_cipd_kernel_package.sh" script (which builds and
uploads a kernel to CIPD) to be better suited for termina_guest instead
of linux_guest.

The main semantic difference here is that we no longer attempt to
package up userspace images in the CIPD package, but only put the kernel
in.

Additionally, we update the CIPD target to avoid using the "linux_guest"
name.

Bug: 84205

Change-Id: Ieca66d45a97a7928968c076bff6af58e6fd07fb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577630
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: David Reveman <reveman@google.com>