[kernel][vm] Pass PageRequest in MapRange
Technically if the SW_FAULT flag is set a PageRequest must be provided
to LookupPagesLocked, although the caller can avoid this if it knows
the target VMO is not pager backed. As such whilst this is not
presently an actual bug, it costs nothing to pass in a PageRequest and
will allow graceful error detection in the future should the current
assumptions get violated.

Change-Id: I0e0686ac6430d4fd3c3ed09c6a2ba60c36bf77ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574810
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>