[core][test] Rotate queues more times for eviction.
Currently there is a single active queue, and so one rotation is
sufficient to move all pages into the evictable state. However, the
number of active queues is changeable and not visible to this test.
To attempt to future proof this test, make it rotate a few times.

Change-Id: I1caf43227b53619d42ee65fb59375908c1081d8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574807
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>