[vulkan] Prototype loading validation layers from package
The validation layers are now included in a component that just serves
them as a directory capability. Applications that want to use the
validation layers can launch layer-server as a component and use the
directory capability. To make everything is hermetic as possible, the
client application can use a specific package hash when launching the
server, ensuring that the validation layer version is the exact same is
the client was compiled against.

Bug: 83307

Change-Id: Ib2c894ed190b28cc00ed4355e5e5fae853e7ba0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574025
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>