[archivist] Change test to use exposed dir
This CL changes the `log_unattributed_stream` test
to connect to LogSink via the exposed directory of
the root of the test Realm instead of its hub path.
This is done because the latter path was causing flaky
behavior for reasons that have yet to be understood.

Bug: 81401
Bug: 81400

Change-Id: Idbd75dfc17abeaa39609cfc0d2262b07d5e0c421
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/578861
Commit-Queue: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>