[kernel][pager] Make PageProvider RefCounted
Changes the PageProvider reference in the PageSource to be a RefPtr,
instead of a unique_ptr. For this commit this is effectively a no-op,
but it unlocks subsequent changes that need additional locations to
hold a reference to the PageProvider.

Change-Id: I0257ce252bdac90b50bababef2d4be6f025f0be7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577626
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>